refactor: centralize php codemod utilities
